requirements.txt
fastapi~=0.103.2
pandas~=2.2.2
pydantic~=2.7.1
redis==5.0.1
requests~=2.32.3
uvicorn~=0.23.2
starlette==0.27.0
python-multipart==0.0.6
websockets
autopep8~=2.3.1
PyYAML~=6.0.1
python-dotenv~=1.0.0
numpy~=1.26.4
chromadb>=0.6.3
langchain>=0.3.18
langchain-chroma>=0.2.2
langchain-community>=0.3.17
langchain-core>=0.3.34
langchain-ollama>=0.2.3
langchain-openai>=0.3.5
langgraph>=0.2.70
pillow>=11.1.0
tiktoken>=0.8.0
rapidfuzz==3.9.2
scikit-learn~=1.3.2
randomname
scalar-fastapi
